requirement.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
configparser==4.0.2
idna==2.9
lxml==4.5.0
numpy==1.16.6
pandas==0.24.2
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==1.9.5
urllib3==1.25.8